feat(appsync): more general mapping template for DynamoDB PutItem
NOTE: More integration tests are missing; looking for early feedback on
the approach.

Currently, the mapping template for DynamoDB PutItem doesn't support
specifying a full primary key, and it doesn't support picking values to
assign to attributes.

This patch introduces a more general approach, which allows setting a
full primary key as well as picking which values from the mutation
argument to assign to table attributes.

It keeps the previous convenient approach of projecting an object to a
set of attribute value assignments via the `Values.project` function.

BREAKING CHANGE: the arguments taken by MappingTemplate.dynamoDbPutItem
have changed.

Fixes #6225

Signed-off-by: Duarte Nunes <duarte@uma.ni>